requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Limiter==3.5.0
google-genai==1.66.0
email-validator==2.3.0
python-dotenv==1.2.1
requests==2.32.5
gunicorn==21.2.0
